requirements.txt
streamlit
tensorflow
numpy
pillow
joblib
opencv-python
lime
scikit-image
matplotlib
